Introduce webfiles() and development web server
The webfiles() rule has been designed primarily for supporting Polymer
web components, but it's general purpose enough to serve the needs of
any web developer. It allows each piece of a website to be defined in
terms of the relationships between source files. This is validated by
checking the src/href/url attributes to make sure the referenced files
exist.  Vulcanization support will also be reinvented shortly.

Each webfiles() rule is executable. It starts a lightweight Jetty web
server that serves up the transitive closure of dependencies with
directory listings and live source reloading.

This rule is capable of serving as a Fileset() replacement, which means
it will be extended in the near future create zip/war archives.