Generate BUILD files for closure-library during loading/analysis phase
This change introduces a new repository rule to generate the BUILD files
for closure-library when the repository is downloaded instead of relying
on checked-in versions (i.e. the files we have in `//closure/library`).

This will eventually require that users migrate their WORKSPACE to
reference the targets directly from the closure library instead of the
ones from this repo. We will provide a buildozer command in the release
notes that can be used to (mostly) do the migration. In the meantime,
we'll update the targets in `//closure/library` to point to the new
targets.