Migrate uses of struct.to_proto
The "to_proto" method on Starlark structs is deprecated and shouldn't be used.
Instead, the proto module's "encode_text" function should be used.
(https://bazel.build/rules/lib/toplevel/proto)

It, along with "to_json", can be disabled in Bazel using the flag
--incompatible_struct_has_no_methods

The underlying implementation is the same, so there should be no observable
changes in final outputs.