Use a private namespace for the java dependencies
The guidelines for rulesets that are not expected to be root
modules (such as this one) is that they only contribute to
the default `maven` workspace if their dependencies are
required by at runtime by production code that uses the
ruleset.

Switching to using a private workspace ensures that this
ruleset plays nicely in larger projects.

Also added a lock file to avoid the need to re-run the maven
dependency resolution each time the bazel daemon starts.